requirements.txt
Flask==2.3.0
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Migrate==4.0.4
email-validator==2.0.0
Werkzeug==2.3.0
Pillow==10.0.0
python-dotenv==1.0.0
WTForms==3.0.1